Changed FragmentSearch.kt (part1)
