+++ app/src/main/java/academy/android/mymovie/clickinterfaces/ContainerListener.kt
@@ -1,10 +1,6 @@
 package academy.android.mymovie.clickinterfaces
 
-import androidx.appcompat.widget.SearchView
-
-
 interface ContainerListener {
-    fun addSearchFragment(searchView: SearchView)
+    fun addSearchFragment()
     fun onSearchClosed()
-    fun closeSearchView()
 }
